add option build

+++ scripts/build.sh
@@ -2,18 +2,22 @@
 
 # Build all
 
+option=$1
+
 path_scripts=./scripts
 deploy_path=$GOPATH/swpc/release
 
-$path_scripts/test-go.sh
-$path_scripts/lint-go.sh
+if [[ $option != 'server' ]]; then
+  $path_scripts/test-go.sh
+  $path_scripts/lint-go.sh
 
-rm -r "$deploy_path/public"
-find "$deploy_path" -maxdepth 1 -type f -delete
+  rm -r "$deploy_path/public"
+  find "$deploy_path" -maxdepth 1 -type f -delete
 
-mkdir -p "$deploy_path/data"
+  mkdir -p "$deploy_path/data"
 
-$path_scripts/build-ui.sh "$deploy_path"
+  $path_scripts/build-ui.sh "$deploy_path"
+fi
 
 GOOS=linux GOARCH=amd64 go build -ldflags="-w -s" -o "$deploy_path/swpc-server" cmd/swpc-server/main.go
